+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### API changes
 * FailedCommandInfo changed so that it stores a WDSession rather than just a Maybe SessionId, thus providing server host and port information as well as the session ID.
-* As a result, mkFailedcommandInfo is now String -> WD FailedCommandInfo, since it requires access to the WDSession state.
+* As a result, mkFailedCommandInfo is now String -> WD FailedCommandInfo, since it requires access to the WDSession state.
 * HTML5StorageType changed to the more accurate WebStorageType
 
 ### new features